[-] [Panel/Outliner] Added Props

+++ Radiant/Include/Radiant/Rendering/Image.hpp
@@ -54,7 +54,7 @@ namespace Radiant
 
 		inline uint32_t CalculateMipCount(uint32_t width, uint32_t height)
 		{
-			return std::floor(std::log2((width < height) ? width : height)) + 1;
+			return (uint32_t)std::floor(std::log2((width < height) ? width : height)) + 1;
 		}
 
 		inline uint32_t GetImageMemorySize(ImageFormat format, uint32_t width, uint32_t height)
